move mapTx test into map_test.go
